independently unpacking layer functions
